feat: 新的背景图 fix: 命令冲突 docs: 添加了命令头和nickname冲突的解释
