+++ liteyuki/liteyuki_main/core.py
@@ -22,9 +22,9 @@ driver = get_driver()
 
 markdown_image = common_db.first(StoredConfig(), default=StoredConfig()).config.get("markdown_image", False)
 
-cmd_liteyuki = on_alconna(
-    Alconna(
-        "liteyuki"
+liteyuki = on_alconna(
+    command=Alconna(
+        "liteecho",
     ),
     permission=SUPERUSER
 )
@@ -73,9 +73,9 @@ switch_image_mode = on_alconna(
 )
 
 
-@cmd_liteyuki.handle()
+@liteyuki.handle()
 async def _(bot: T_Bot):
-    await cmd_liteyuki.finish(f"Hello, Liteyuki!\nBot {bot.self_id}\nLiteyukiID {config.get('liteyuki_id', 'No')}")
+    await liteyuki.finish(f"Hello, Liteyuki!\nBot {bot.self_id}")
 
 
 @update_liteyuki.handle()
